Add a project brief in LaTeX
Add the brief as required at the beginning of the project. It is written in
LaTeX (Yes, I've started to learn it!). The brief is in 3 parts, the problem,
the goals, and the scope for expansion. This is then grouped together in the
overall brief. In order to create the PDF that is required for handin, a
Makefile is also included.
This is essentially the first draft of the brief, as is a complete draft.