Modify ring buffer to no longer contain a position pointer
Remove the position pointer, as the ring buffer is likely to be filled and read
by two separate functions, as such it would need to contain at least 2 position
pointers to be able to function, and to increment or decrement the pointer you'd
need to either specify which position to adjust, or have different functions
for each pointer.
Now the position pointer is kept completely separate, by the function that's
accessing the buffer, and increment and decrement now take the pointer to change
as an argument.
Also the ring buffer no longer contains the actual buffer, just a pointer to it.

git-svn-id: svn+ssh://svn.forge.ecs.soton.ac.uk/projects/sanchp3yp@22 bc9da1e1-76ae-0410-a611-e94dc21032b9