Add a Matlab function that achieves cross-correlation manually
Returns an array giving a good approximation to the cross-correlation of two
arrays. Is achieved in a manner that can be easily ported to C.

Includes 3 different methods, 2 of which are commented out, either due to
producing errors when running, or due to taking far too long to complete. While
not optimal, the last (uncommented out) method will function in a short length
of time (currently ~0.3 seconds for arrays of 5000 samples and max delay of
+/- 200)

git-svn-id: svn+ssh://svn.forge.ecs.soton.ac.uk/projects/sanchp3yp@83 bc9da1e1-76ae-0410-a611-e94dc21032b9