Add a function to generate the error value for the LMS
This will definitely be subsumed into the rest of the code later on
NOTE: It does an addition when it would normally be expected to do a
subtraction. This is to cause the estimated signal produced by the LMS filter
to be the inverse of the incoming signal, reducing complexity later on as I
do not need to invert this separately manually, reducing computational
complexity.